+++ packages/server/src/server/handlers.ts
@@ -5,7 +5,7 @@ export * as memory from './handlers/memory';
 export * as network from './handlers/network';
 export * as telemetry from './handlers/telemetry';
 export * as tools from './handlers/tools';
-export * as workflows from './handlers/workflows';
-export * as vNextWorkflows from './handlers/vNextWorkflows';
+export * as workflows from './handlers/legacyWorkflows';
+export * as vNextWorkflows from './handlers/workflows';
 export * as vector from './handlers/vector';
 export * as voice from './handlers/voice';
